extract notrandombytes into a separate file
